Extraire liveData du DB dataGrid
