+++ ponysdk/src/main/java/com/ponysdk/core/ui/datagrid2/SimpleDataSource.java
@@ -24,7 +24,6 @@
 package com.ponysdk.core.ui.datagrid2;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -33,46 +32,29 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Objects;
 import java.util.Set;
-import java.util.function.Consumer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.ponysdk.core.ui.datagrid2.SimpleDataGridController.Column;
 import com.ponysdk.core.ui.datagrid2.SimpleDataGridController.ColumnControllerSort;
-import com.ponysdk.core.ui.datagrid2.SimpleDataGridController.Interval;
+import com.ponysdk.core.ui.datagrid2.SimpleDataGridController.RenderingHelpersCache;
 import com.ponysdk.core.ui.datagrid2.SimpleDataGridController.Row;
 
 public abstract class SimpleDataSource<K, V> implements DataGridSource<K, V> {
 
     protected static final Logger log = LoggerFactory.getLogger(SimpleDataSource.class);
-    protected final Map<K, Row<V>> cache = new HashMap<>();
     protected final LinkedHashMap<Object, Comparator<Row<V>>> sorts = new LinkedHashMap<>();
     protected final Map<Object, AbstractFilter<V>> filters = new HashMap<>();
-    protected final List<Row<V>> liveData = new ArrayList<>();
     protected final List<Row<V>> liveSelectedData = new ArrayList<>();
     protected final Set<K> selectedKeys = new HashSet<>();
     protected DataGridAdapter<K, V> adapter;
+    protected RenderingHelpersCache<V> renderingHelpersCache; //FIXME : listener de dataSource
     protected int rowCounter = 0;
 
     //----------------------------------------------------------------------------------------------------------//
-    //----------------------------------------- Cache getters --------------------------------------------------//
-    //----------------------------------------------------------------------------------------------------------//
-
-    @Override
-    public Row<V> getRow(final K k) {
-        return cache.get(k);
-    }
-
-    @Override
-    public Collection<Row<V>> getRows() {
-        return cache.values();
-    }
-
-    //----------------------------------------------------------------------------------------------------------//
-    //------------------------------------------------ Adapter -------------------------------------------------//
+    //---------------------------------------- Adapter / RenderingCache ----------------------------------------//
     //----------------------------------------------------------------------------------------------------------//
 
     @Override
@@ -80,150 +62,15 @@ public abstract class SimpleDataSource<K, V> implements DataGridSource<K, V> {
         this.adapter = adapter;
     }
 
-    //----------------------------------------------------------------------------------------------------------//
-    //----------------------------------------- Gestion de liveData --------------------------------------------//
-    //----------------------------------------------------------------------------------------------------------//
-
-    // Insert data in the cache or update it if it already exists
     @Override
-    public Interval setData(final V v) {
-        Objects.requireNonNull(v);
-        final K k = adapter.getKey(v);
-        final Row<V> row = cache.get(k);
-        Interval interval;
-        if (row != null) {
-            if (row.data == v) return null;
-            interval = updateData(k, row, v);
-        } else {
-            interval = insertData(k, v);
-        }
-        return interval;
-    }
-
-    // Rows are updated when they already exist
-    private Interval updateData(final K k, final Row<V> row, final V newV) {
-        if (row.accepted) {
-            final int oldLiveDataSize = liveData.size();
-            final int oldRowIndex = removeRow(liveData, row);
-            final boolean selected = selectedKeys.contains(k);
-            if (selected) removeRow(liveSelectedData, row);
-            row.data = newV;
-            return onWasAcceptedAndRemoved(selected, row, oldLiveDataSize, oldRowIndex);
-        } else {
-            row.data = newV;
-            return onWasNotAccepted(k, row);
-        }
+    public void sort() {
+        liveSelectedData.sort(this::compare);
     }
 
     @Override
-    public Interval updateData(final K k, final Consumer<V> updater) {
-        final Row<V> row = cache.get(k);
-        if (row == null) return null;
-        if (row.accepted) {
-            final int oldLiveDataSize = liveData.size();
-            final int oldRowIndex = removeRow(liveData, row);
-            final boolean selected = selectedKeys.contains(k);
-            if (selected) removeRow(liveSelectedData, row);
-            updater.accept(row.data);
-            return onWasAcceptedAndRemoved(selected, row, oldLiveDataSize, oldRowIndex);
-        } else {
-            updater.accept(row.data);
-            return onWasNotAccepted(k, row);
-        }
+    public void setRenderingHelpersCache(final RenderingHelpersCache<V> renderingHelpersCache) {
+        this.renderingHelpersCache = renderingHelpersCache;
     }
-
-    @Override
-    public V removeData(final K k) {
-
-        final Row<V> row = cache.remove(k);
-        final boolean selected = selectedKeys.remove(k);
-        if (row.accepted) {
-            //            final int oldLiveDataSize = liveData.size();
-            //            final int rowIndex = removeRow(liveData, row);
-            if (selected) {
-                removeRow(liveSelectedData, row);
-            }
-            //            refreshRows(rowIndex, oldLiveDataSize);
-        }
-        return row.data;
-    }
-
-    // Here rows are created and inserted in liveData
-    private Interval insertData(final K k, final V data) {
-        final Row<V> row = new Row<>(rowCounter++, data);
-        row.accepted = accept(row);
-        cache.put(k, row);
-        if (!row.accepted) return null;
-        final int rowIndex = insertRow(liveData, row);
-        return new Interval(rowIndex, liveData.size());
-    }
-
-    private Interval onWasAcceptedAndRemoved(final boolean selected, final Row<V> row, final int oldLiveDataSize,
-                                             final int oldRowIndex) {
-        //        clearRenderingHelpers(row);
-        if (accept(row)) {
-            final int rowIndex = insertRow(liveData, row);
-            if (selected) insertRow(liveSelectedData, row);
-            if (oldRowIndex <= rowIndex) {
-                return new Interval(oldRowIndex, rowIndex + 1);
-            } else {
-                return new Interval(rowIndex, oldRowIndex + 1);
-            }
-        } else {
-            row.accepted = false;
-            return new Interval(oldRowIndex, oldLiveDataSize);
-        }
-    }
-
-    private Interval onWasNotAccepted(final K k, final Row<V> row) {
-        //        clearRenderingHelpers(row);
-        if (accept(row)) {
-            row.accepted = true;
-            final int rowIndex = insertRow(liveData, row);
-            if (selectedKeys.contains(k)) insertRow(liveSelectedData, row);
-            return new Interval(rowIndex, liveData.size());
-        } //else do nothing
-        return null;
-    }
-
-    private boolean accept(final Row<V> row) {
-        for (final AbstractFilter<V> filter : filters.values()) {
-            if (!filter.test(row)) return false;
-        }
-        return true;
-    }
-
-    @Override
-    public void resetLiveData() {
-        liveSelectedData.clear();
-        liveData.clear();
-        for (final Row<V> row : cache.values()) {
-            row.accepted = accept(row);
-            if (row.accepted) {
-                insertRow(liveData, row);
-                if (selectedKeys.contains(adapter.getKey(row.data))) {
-                    insertRow(liveSelectedData, row);
-                }
-            }
-        }
-    }
-
-    //    public void refreshRows(final int from, final int to) {
-    //        this.from = Math.min(this.from, from);
-    //        this.to = Math.max(this.to, to);
-    //        if (SimpleDataGridController.bound) doRefreshRows(); //FIXME
-    //    }
-    //
-    //    public void doRefreshRows() {
-    //        try {
-    //            if (listener != null) {
-    //                listener.onUpdateRows(from, to);
-    //            }
-    //        } finally {
-    //            from = Integer.MAX_VALUE;
-    //            to = 0;
-    //        }
-    //    }
 
     //----------------------------------------------------------------------------------------------------------//
     //------------------------------- BinarySearch insertion, removal, findIndex -------------------------------//
@@ -278,7 +125,7 @@ public abstract class SimpleDataSource<K, V> implements DataGridSource<K, V> {
 
     protected final int compare(final Row<V> r1, final Row<V> r2) {
 
-        for (final Comparator<Row<V>> sort : getSorts()) {
+        for (final Comparator<Row<V>> sort : sorts.values()) {
             final int diff = sort.compare(r1, r2);
             if (diff != 0) return diff;
         }
@@ -322,10 +169,10 @@ public abstract class SimpleDataSource<K, V> implements DataGridSource<K, V> {
         sort();
     }
 
-    @Override
-    public Collection<Comparator<Row<V>>> getSorts() {
-        return sorts.values();
-    }
+    //    @Override
+    //    public Collection<Comparator<Row<V>>> getSorts() {
+    //        return sorts.values();
+    //    }
 
     @Override
     public Set<Entry<Object, Comparator<Row<V>>>> getSortsEntry() {
@@ -375,17 +222,16 @@ public abstract class SimpleDataSource<K, V> implements DataGridSource<K, V> {
         return selectedKeys.contains(k);
     }
 
-    @Override
-    public void select(final K k) {
-        final Row<V> row = cache.get(k);
-        if (row == null || !selectedKeys.add(k) || !row.accepted) return;
-        insertRow(liveSelectedData, row);
+    protected boolean accept(final Row<V> row) {
+        for (final AbstractFilter<V> filter : filters.values()) {
+            if (!filter.test(row)) return false;
+        }
+        return true;
     }
 
     @Override
-    public void unselect(final K k) {
-        final Row<V> row = cache.get(k);
-        if (row == null || !selectedKeys.remove(k) || !row.accepted) return;
-        removeRow(liveSelectedData, row);
+    public void unselectAllData() {
+        liveSelectedData.clear();
+        selectedKeys.clear();
     }
 }
